big cards refactor for better maintainability
